+++ lumenairy/analysis/__init__.py
@@ -26,6 +26,8 @@ from .analysis import (
     beam_d4sigma,
     beam_power,
     strehl_ratio,
+    coupling_efficiency,
+    M2,
     check_sampling_conditions,
     compute_psf,
     compute_otf,
@@ -111,6 +113,7 @@ from .plotting import (
 __all__ = [
     # analysis
     'beam_centroid', 'beam_d4sigma', 'beam_power', 'strehl_ratio',
+    'coupling_efficiency', 'M2',
     'check_sampling_conditions', 'compute_psf', 'compute_otf',
     'compute_mtf', 'mtf_radial', 'remove_wavefront_modes',
     'opd_pv_rms', 'wave_opd_1d', 'wave_opd_2d', 'check_opd_sampling',